[feat] Order 사장 주문 상세 보기 추가

+++ src/main/java/com/sparta/spartdelivery/entity/OrderDetail.java
@@ -22,6 +22,6 @@ public class OrderDetail{
     @JoinColumn(name = "menu_id", referencedColumnName = "menuId", nullable = false)
     private Menu menu;
 
-    private int quantity;
+    private Short quantity;
 
 }
